requirements.txt
asgiref==3.8.1
Django==5.1.1
django-cors-headers==4.5.0
django-filter==24.3
djangorestframework==3.15.2
psycopg2==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.1
tzdata==2024.2
